Provide "length" parameter in reading of replies
Provide the "length" parameter in the reply-reading functions, which is
required by e.g. GetImage and GetKeyboardMapping. The length parameter
has to be provided as is, i.e. not multiplied by 4, since the
interpretation of the parameter is not constant across all replies. This
also means that the size of a reply cannot (necessarily) be derived from
"length". The required number of padding needed to be read has as a
result been fixed.