Make to-frame/to-value return raw frame/value
Make to-frame/to-value return raw frame/value, i.e. content of request
without opcode, length, or extra padding. The generation of complete
frames/values is now the responsibility of send, which now resides in
conn-internal instead of gen-common.

This change is made as to decrease the amount of code in xmlgen and to
simplify the implementation of extensions in the future, which contains
complications such as major/minor opcodes.