Fix detection of primitive types
The old typemap combines primitive and composite types together so that
it's impossible to know whether a type specified in XML is a primitive.
The new typemap separates the primitive and composite types.

Checking of primitive types was only checking newly defined primitives
in the current XML file, causing any references to primitives in other
files to be mis-categorized as composite. Now both newly defined
primitives (in PRIMITIVE-TYPES) and primitives from outside (in TYPEMAP)
are checked.