Make skipping of replies separate from requests
Since the XML definition of <reply> elements are always inside
<request>, it used to be that when a request has to be skipped, the
reply inside will also be skipped. This results in having to manually
create reply functions and types for e.g. QueryTextExtents. Now the
skipping of replies are separate from the skipping of requests, so
perfectly parse-able replies don't have to be manually written.