Simplify connection setup
Since we now have the setup structs as data, it is now used to
deserialize the reply returned by server during setup. Rather than
hand-coding and reading the reply field by field, we now do block reads
and pass the resulting buffer to the setup structs directly to be
deserialized.