Update read-channel loop for gen-common
Update read-channel loop for gen-common. Now we do block reads and pass
the resulting buffer directly to respective read-reply/event/error.

It used to be that the first field in a reply is assumed to be unsigned,
since it has to be read in the read-channel loop (we need the sequence
number, which comes after the first field), and there's no way to
re-read it in read-reply as we're directly reading from channel. Now
that we have the buffer, we no longer need the unsigned assumption and
read-reply can read the first field properly.